+++ src/events/guildDelete.js
@@ -2,7 +2,7 @@
 import {  Events, ChannelType, PermissionsBitField  } from "discord.js";
 import global from "../utils/global.js";
 import guildManager from "../utils/guildManager.js";
-import logtail from "../utils/logtail.js";
+import logtail from "../utils/logs/logtail.js";
 import {  DateTime  } from "luxon";
 
 
@@ -11,16 +11,17 @@ import {  DateTime  } from "luxon";
 // Event Execution:
 export default {
     name: Events.GuildDelete,
+    /** @param {import('discord.js').Guild} guild */
     async execute(guild) {
         // Debug - Leaving Guild:
         if(global.outputDebug_InDepth) {
             console.log('guildDelete Event Fired!:')
             console.log(`guildID: ${guild.id}`)
         }
-        const joinedAtDateString = DateTime.fromSeconds(guild?.joinedTimestamp).toLocaleString(DateTime.DATETIME_FULL)
+        const joinedAtDateString = DateTime.fromMillis(guild?.joinedTimestamp).toLocaleString(DateTime.DATETIME_FULL)
         logtail.info(`Guild ${guild?.id} has removed Sessions Bot!`, {guildName: guild?.name, memberCount: guild?.memberCount, joinedAt: joinedAtDateString});
 
         // Move guild to archive within database:
-        await guildManager.guilds(guild.id).archiveGuild();
+        await guildManager.guilds(guild.id).archiveGuild(guild);
     }
 }