+++ src/utils/guildManager.js
@@ -13,10 +13,9 @@ import { // Discord.js:
     MessageFlags,
     SectionBuilder,
     ThreadAutoArchiveDuration,
-    ChannelType,
-    ThumbnailBuilder
+    ChannelType
 } from 'discord.js';
-import logtail from "../utils/logtail.js";
+import logtail from "./logs/logtail.js";
 
 const inDepthDebug = (c) => { if (global.outputDebug_InDepth) { console.log(`[Guild Manager]: ${c}`) } }
 
@@ -48,6 +47,16 @@ const guilds = (guildId) => {return {
             // Save new guild to database:
             await db.collection('guilds').doc(String(guildId)).set(defaultGuildData, { merge: true });
 
+            // Save guild join log:
+            const guildBotData = await global.client.guilds.fetch(guildId)
+            await db.collection('events').doc('inviteLogs').collection('guilds').doc(String(guildId)).set({
+                guildId: guildBotData?.id,
+                guildName: guildBotData?.name,
+                guildDesc: guildBotData?.description,
+                memberCount: guildBotData?.memberCount,
+                joinedAt: new Date()
+            }, { merge: true });
+
             // Success:
             console.log(`[+] Successfully added new guild! Id: ${guildId}`);
             const result = { success: true, data: 'Successfully added new guild!' };
@@ -78,21 +87,18 @@ const guilds = (guildId) => {return {
 
 
     // Move Guild to Archive:
-    archiveGuild: async () => {
+    archiveGuild: async (guildBotData) => {
         const guildRef = db.collection('guilds').doc(guildId);
         const archivedRef = db.collection('archivedGuilds').doc(guildId);
 
         try {
             // 1. Read original doc
             const guildDoc = await guildRef.get();
-
             if (!guildDoc.exists) {
-
                 console.warn(`Guild document ${guildId} does not exist. Failed to archive!`);
-
+                logtail.warn(`Guild document ${guildId} does not exist. Failed to archive!`);
                 return { success: false, error: `Couldn't find existing guild doc to archive!` };
             }
-
             const guildData = guildDoc.data();
 
             // 2. Write to archive
@@ -101,7 +107,19 @@ const guilds = (guildId) => {return {
                 archivedAt: new Date()
             });
 
-            // 3. Delete original
+            // 3. Save guild leave log:
+            const joinedAtDateString = DateTime.fromMillis(guildBotData?.joinedTimestamp).toLocaleString(DateTime.DATETIME_FULL);
+            const removedAtDateString = DateTime.now().toLocaleString(DateTime.DATETIME_FULL);
+            await db.collection('events').doc('removeLogs').collection('guilds').doc(String(guildId)).set({
+                guildId: guildBotData?.id,
+                guildName: guildBotData?.name,
+                guildDesc: guildBotData?.description,
+                memberCount: guildBotData?.memberCount,
+                joinedAt: joinedAtDateString,
+                removedAt: removedAtDateString
+            }, { merge: true });
+
+            // 4. Delete original
             await guildRef.delete();
 
             console.log(`[-] Guild ${guildId} moved to archive successfully.`);
@@ -142,11 +160,6 @@ const guilds = (guildId) => {return {
 
 // Guild Configuration - Nested Functions:
 const guildConfiguration = (guildId) => {return {
-
-    // !!! [NEEDS COMPLETION] - {guildConfiguration}
-    //{*} Remove Specific Session Schedule:
-        // removeSessionSchedule: async (sessionId) => {},
-    // !!!
 
     // -- Top Level Configure Function:
     configureGuild : async (configuration) => {
@@ -412,7 +425,7 @@ const guildPanel = (guildId) => {return {
         const sessionSignup = async () => {
             
             // Append Each Upcoming Session:
-            for ([sessionId, sessionData] of upcomingSessions) {
+            for (const [sessionId, sessionData] of upcomingSessions) {
                 const contentAttempt = await guildSessions(guildId).getSessionPanelContents(sessionId, sessionData, accentColor);
                 if(contentAttempt.success){
                     // Add this Session to Components: